+++ borax/calendars/festivals2.py
@@ -30,6 +30,8 @@ class FreqConst:
     YEARLY = 0
     MONTHLY = 1
 
+    LABEL2VAL = {'monthly': 1, 'yearly': 0, 'm': 1, 'y': 0}
+
 
 class FestivalCatalog:
     basic = 'basic'
@@ -167,8 +169,9 @@ class Period:
     """A shortcut methods for some specified date Period."""
 
     @staticmethod
-    def solar_year(year: int) -> Tuple[date, date]:
-        return date(year, 1, 1), date(year, 12, 31)
+    def solar_year(year: int, end_year: int = 0) -> Tuple[date, date]:
+        end_year = end_year or year
+        return date(year, 1, 1), date(end_year, 12, 31)
 
     @staticmethod
     def solar_month(year: int, month: int) -> Tuple[date, date]:
@@ -176,8 +179,9 @@ class Period:
         return date(year, month, 1), date(year, month, ndays)
 
     @staticmethod
-    def lunar_year(year: int) -> Tuple[LunarDate, LunarDate]:
-        return LunarDate(year, 1, 1), LunarDate(year + 1, 1, 1) - timedelta(days=1)
+    def lunar_year(year: int, end_year: int = 0) -> Tuple[LunarDate, LunarDate]:
+        end_year = end_year or year
+        return LunarDate(year, 1, 1), LunarDate.last_day(end_year)
 
     @staticmethod
     def lunar_month(year: int, month: int, leap: int = _IGNORE_LEAP_MONTH) -> Tuple[LunarDate, LunarDate]:
@@ -460,12 +464,17 @@ class SolarFestival(Festival):
     """
     date_class = date
 
-    def __init__(self, *, day: int, freq: int = FreqConst.YEARLY, month: int = 0, name: str = None):
+    def __init__(self, *, day: int, freq: Union[int, Literal['yearly', 'monthly', 'y', 'm']] = FreqConst.YEARLY,
+                 month: int = 0, name: str = None):
         if day < 0:
             day = -day
             reverse = 1
         else:
             reverse = 0
+        if isinstance(freq, str):
+            freq = FreqConst.LABEL2VAL.get(freq, -1)
+            if freq == -1:
+                raise ValueError('Invalid freq string.')
         super().__init__(name=name, freq=freq, month=month, day=day, reverse=reverse, schema=FestivalSchema.SOLAR)
 
     def _get_description(self) -> str:
@@ -680,13 +689,17 @@ class LunarFestival(Festival):
     """
     date_class = LunarDate
 
-    def __init__(self, *, day: int, freq: int = FreqConst.YEARLY, month: int = 0, leap: int = _IGNORE_LEAP_MONTH,
-                 name: str = None):
+    def __init__(self, *, day: int, freq: Union[int, Literal['yearly', 'monthly', 'y', 'm']] = FreqConst.YEARLY,
+                 month: int = 0, leap: int = _IGNORE_LEAP_MONTH, name: str = None):
         if day < 0:
             day = -day
             reverse = 1
         else:
             reverse = 0
+        if isinstance(freq, str):
+            freq = FreqConst.LABEL2VAL.get(freq, -1)
+            if freq == -1:
+                raise ValueError('Invalid freq string.')
         super().__init__(freq=freq, name=name, month=month, day=day, leap=leap, reverse=reverse,
                          schema=FestivalSchema.LUNAR)