+++ borax/calendars/lunardate.py
@@ -659,6 +659,24 @@ class LunarDate:
         return cls.from_solar_date(sd.year, sd.month, sd.day)
 
     @classmethod
+    def last_day(cls, year: int, month: int = 0, leap: int = 0) -> 'LunarDate':
+        """return the last day in a lunar year or a lunar month."""
+        mdls = list(LCalendars.iter_year_month(year))
+        if month == 0:
+            index = -1
+        else:
+            leap_month_of_year = LCalendars.leap_month(year)
+            if leap:
+                if month == leap_month_of_year:
+                    index = month
+                else:
+                    raise InvalidLunarDateError(f'{year=}, {month=},leap=1 does not exist.')
+            else:
+                index = month - 1 + int(leap_month_of_year and month > leap_month_of_year)
+        month, day, leap = mdls[index]
+        return cls(year, month, day, leap)
+
+    @classmethod
     def strptime(cls, date_str: str, date_fmt: str) -> 'LunarDate':
         """Parse a LunarDate object from a whole string.
 
@@ -678,7 +696,7 @@ class LunarDate:
         :param other: a instance of LunarDate / date / timedelta
         :return:
         """
-        if hasattr(other, 'solar'):
+        if hasattr(other, 'solar'):  # For WrappedDate in festivals2 module
             return self.to_solar_date() - other.solar
         elif isinstance(other, LunarDate):
             return self.to_solar_date() - other.to_solar_date()